Fix Next Button Bug
Fixed the Next Button Bug such that now clicking the True or False
Buttons when the question prompt says, "Click Next Button" does not
lead to a Correct or Incorrect Response. Also added bringing up
the next question immediately after a response to a question
and not having to click Next to advance to the next question.